requirements.txt
numpy 
opencv-python 
pillow 
pytesseract 
PyWavelets
matplotlib
scikit-image